+++ UD1/src/ejercicios/ej07.java
@@ -5,10 +5,10 @@ import java.util.Scanner;
 public class ej07 {
 	public static void main(String[] args) {
 		Scanner s = new Scanner(System.in);
-		int euros = 1;
+		double euros = 1;
 		double dollar;
 		System.out.println("Introduce la cantidad de dollares a conversar: ");
-		dollar = s.nextInt();
+		dollar = s.nextDouble();
 		
 		double calculo = euros * dollar;
 		System.out.println("El cambio es de: " + calculo + "$");
